% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttMOMOCut.R
\name{AttMOMOCut}
\alias{AttMOMOCut}
\title{Run AttMOMO base and baseline adjusted models.
Read input data, create output directories and return csv-file with estimates.
Demanded input files must be available in wdir/data}
\usage{
AttMOMOCut(
  country,
  wdir,
  StartWeek,
  EndWeek,
  groups,
  pooled = NULL,
  indicators,
  indicatorCuts,
  lags = 2,
  ptrend = 0.05,
  p26 = 0.05,
  p52 = 0.1,
  Rdata = FALSE
)
}
\arguments{
\item{country}{Country name.}

\item{wdir}{Working directory.}

\item{StartWeek}{ISOweek format}

\item{EndWeek}{ISOweek format}

\item{groups}{list of group names}

\item{pooled}{list of group names to be pooled (default = NULL)
Must be part of groups.}

\item{indicators}{list if indicator variables.
One file for each must be available: IndicatorName_data.txt}

\item{indicatorCuts}{list of cut-weeks for each indicator in indicators}

\item{lags}{weeks of lagged effect (default = 2, max = 9)}

\item{ptrend}{significance of trend to be included (default = 0.05)}

\item{p26}{significance of half year-sine be included (default = 0.05)}

\item{p52}{significance of year-sine be included (default = 0.10)}

\item{Rdata}{Return data to R (TRUE/FALSE - default FALSE).}
}
\value{
Write a ;-separated file AttData_indicators.txt in output directory/output.
}
\description{
Run AttMOMO base and baseline adjusted models.
Read input data, create output directories and return csv-file with estimates.
Demanded input files must be available in wdir/data
}
