% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetET.R
\name{GetET}
\alias{GetET}
\title{Excess Temperature data from weather data}
\usage{
GetET(ET, StartWeek, EndWeek, tvar)
}
\arguments{
\item{ET}{Data containing the variables: date, pop3, NUTS3, tvar}

\item{StartWeek}{Start week (ISO format)}

\item{EndWeek}{End week (ISO format)}

\item{tvar}{name of temperature variable in data}
}
\value{
data.table with ISOweek and ET (excess temperature)
}
\description{
Excess Temperature data from weather data
}
