% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMOMOdata.R
\name{GetMOMOdata}
\alias{GetMOMOdata}
\title{Convert an A-MOMO complete file to an AttMOMO input file}
\usage{
GetMOMOdata(indir, filename, outdir = NULL, outfile = TRUE)
}
\arguments{
\item{indir}{Full path to the directory where A-MOMO complete file is placed.}

\item{filename}{Name of A-MOMO complete file incl. extension (.txt)}

\item{outdir}{Directory where the AttMOMO input file should be placed (default indir)}

\item{outfile}{Should output be written an AttMOMO deaths input ;-separated file, death_data.txt, in outdir (default TRUE)}
}
\value{
AttMOMO input file: group, ISOweek, deaths
}
\description{
Convert an A-MOMO complete file to an AttMOMO input file
}
