% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetWdata.R
\name{GetWdata}
\alias{GetWdata}
\title{Read weather data achieved from EuroMOMO}
\usage{
GetWdata(dir, CountryCode, NUTS = NULL)
}
\arguments{
\item{dir}{Directory with data: dir/wdata_CountryCode.txt.}

\item{CountryCode}{NUTS country code.}

\item{NUTS}{A list of equal length NUTS by which to limit data (default = CountryCode).}
}
\value{
data.table: date, pop3, NUTS3, temp, mintemp, maxtemp.
}
\description{
Read weather data achieved from EuroMOMO
}
